Refactor supported instruments profile construction
