+++ domain/src/main/java/com/alligator/market/domain/provider/handler/AbstractInstrumentHandler.java
@@ -1,8 +1,6 @@
 package com.alligator.market.domain.provider.handler;
 
 import com.alligator.market.domain.instrument.Instrument;
-import com.alligator.market.domain.instrument.classification.Asset;
-import com.alligator.market.domain.instrument.classification.Product;
 import com.alligator.market.domain.instrument.vo.InstrumentCode;
 import com.alligator.market.domain.marketdata.tick.model.QuoteTick;
 import com.alligator.market.domain.provider.MarketDataProvider;
@@ -10,7 +8,6 @@ import com.alligator.market.domain.provider.handler.instrument.SupportedInstrume
 import com.alligator.market.domain.provider.vo.HandlerCode;
 import org.reactivestreams.Publisher;
 
-import java.util.LinkedHashSet;
 import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.atomic.AtomicReference;
@@ -33,9 +30,8 @@ public abstract class AbstractInstrumentHandler<P extends MarketDataProvider, I 
     /**
      * Конструктор: обработчик сам выводит свой профиль из набора поддерживаемых инструментов.
      * <ul>
-     *     <li>Выполняет базовые проверки параметров;</li>
      *     <li>Собирает сводный профиль инструментов, поддерживаемых обработчиком;</li>
-     *     <li>Валидирует инварианты: java-класс инструмента, единые asset/product и уникальность instrumentCode;</li>
+     *     <li>Передает валидацию и построение профиля в {@link SupportedInstrumentsProfile};</li>
      * </ul>
      */
     protected AbstractInstrumentHandler(
@@ -43,17 +39,11 @@ public abstract class AbstractInstrumentHandler<P extends MarketDataProvider, I 
             Class<I> instrumentClass,
             Set<? extends I> supportedInstruments
     ) {
-        Objects.requireNonNull(handlerCode, "handlerCode must not be null");
-        Objects.requireNonNull(instrumentClass, "instrumentClass must not be null");
-        Objects.requireNonNull(supportedInstruments, "supportedInstruments must not be null");
-
-        if (supportedInstruments.isEmpty()) {
-            throw new IllegalArgumentException("supportedInstruments must not be empty");
-        }
-
-        SupportedInstrumentsProfile profile =
-                buildSupportedInstrumentProfile(handlerCode, instrumentClass, supportedInstruments);
-
+        SupportedInstrumentsProfile profile = SupportedInstrumentsProfile.fromSupportedInstruments(
+                handlerCode,
+                instrumentClass,
+                supportedInstruments
+        );
         this.handlerCode = handlerCode;
         this.supportedInstrumentsProfile = profile;
     }
@@ -102,93 +92,6 @@ public abstract class AbstractInstrumentHandler<P extends MarketDataProvider, I 
      */
     protected final P provider() {
         return requireAttachedProvider();
-    }
-
-    /*
-     * Собирает и валидирует сводный профиль поддерживаемых инструментов.
-     */
-    private static <I extends Instrument> SupportedInstrumentsProfile buildSupportedInstrumentProfile(
-            HandlerCode handlerCode,
-            Class<I> instrumentClass,
-            Set<? extends I> supportedInstruments
-    ) {
-        Asset resolvedAsset = null;
-        Product resolvedProduct = null;
-        LinkedHashSet<InstrumentCode> resolvedCodes = new LinkedHashSet<>();
-
-        for (I instrument : supportedInstruments) {
-            Objects.requireNonNull(instrument, "supportedInstrument must not be null");
-
-            InstrumentCode instrumentCode = Objects.requireNonNull(
-                    instrument.instrumentCode(),
-                    "supported instrumentCode must not be null"
-            );
-
-            if (!instrumentClass.isInstance(instrument)) {
-                throw new IllegalArgumentException(
-                        "Supported instrument '%s' has java class '%s', but handler '%s' expects '%s'"
-                                .formatted(
-                                        instrumentCode.value(),
-                                        instrument.getClass().getName(),
-                                        handlerCode.value(),
-                                        instrumentClass.getName()
-                                )
-                );
-            }
-
-            Asset currentAsset = Objects.requireNonNull(
-                    instrument.asset(),
-                    "supported asset must not be null"
-            );
-
-            Product currentProduct = Objects.requireNonNull(
-                    instrument.product(),
-                    "supported product must not be null"
-            );
-
-            if (resolvedAsset == null) {
-                resolvedAsset = currentAsset;
-                resolvedProduct = currentProduct;
-            } else {
-                if (currentAsset != resolvedAsset) {
-                    throw new IllegalArgumentException(
-                            "Supported instruments of handler '%s' have inconsistent asset: expected '%s', but found '%s' for instrument '%s'"
-                                    .formatted(
-                                            handlerCode.value(),
-                                            resolvedAsset,
-                                            currentAsset,
-                                            instrumentCode.value()
-                                    )
-                    );
-                }
-
-                if (currentProduct != resolvedProduct) {
-                    throw new IllegalArgumentException(
-                            "Supported instruments of handler '%s' have inconsistent product: expected '%s', but found '%s' for instrument '%s'"
-                                    .formatted(
-                                            handlerCode.value(),
-                                            resolvedProduct,
-                                            currentProduct,
-                                            instrumentCode.value()
-                                    )
-                    );
-                }
-            }
-
-            if (!resolvedCodes.add(instrumentCode)) {
-                throw new IllegalArgumentException(
-                        "Supported instrument code '%s' is duplicated in handler '%s'"
-                                .formatted(instrumentCode.value(), handlerCode.value())
-                );
-            }
-        }
-
-        return new SupportedInstrumentsProfile(
-                freezeSupportedInstrumentCodes(resolvedCodes),
-                instrumentClass,
-                resolvedAsset,
-                resolvedProduct
-        );
     }
 
     /*
@@ -261,18 +164,4 @@ public abstract class AbstractInstrumentHandler<P extends MarketDataProvider, I 
         );
     }
 
-    /*
-     * Возвращает неизменяемую защищенную копию списка поддерживаемых инструментов.
-     */
-    private static Set<InstrumentCode> freezeSupportedInstrumentCodes(Set<InstrumentCode> supportedInstrumentCodes) {
-        Objects.requireNonNull(supportedInstrumentCodes, "supportedInstrumentCodes must not be null");
-
-        LinkedHashSet<InstrumentCode> copy = new LinkedHashSet<>(supportedInstrumentCodes.size());
-        for (InstrumentCode code : supportedInstrumentCodes) {
-            Objects.requireNonNull(code, "code must not be null");
-            copy.add(code);
-        }
-
-        return java.util.Collections.unmodifiableSet(copy);
-    }
 }
